Update code-native integration and custom connector boilerplate
