+++ src/commands/integrations/init/index.ts
@@ -1,7 +1,12 @@
-import { Command, Args } from "@oclif/core";
+import path from "path";
+import { Args, Command } from "@oclif/core";
 import fs from "fs/promises";
+import { camelCase } from "lodash-es";
+import { v4 as uuid4 } from "uuid";
+import { prismaticUrl } from "../../../auth.js";
+import { template, updatePackageJson } from "../../../generate/util.js";
 import { VALID_NAME_REGEX } from "../../../utils/generate.js";
-import GenerateIntegrationCommand from "./integration.js";
+import { devDependencies } from "../../../utils/devDependencies.js";
 
 export default class InitializeIntegration extends Command {
   static description = "Initialize a new Code Native Integration";
@@ -30,12 +35,79 @@ export default class InitializeIntegration extends Command {
     await fs.mkdir(name);
     process.chdir(name);
 
-    await GenerateIntegrationCommand.invoke(
-      {
-        name,
-        description: "Prism-generated Integration",
+    const context = {
+      integration: { name, description: "Prism-generated Integration", key: camelCase(name) },
+      flow: {
+        stableKey: uuid4(),
       },
-      this.config,
-    );
+      configVars: {
+        connection: {
+          stableKey: uuid4(),
+        },
+        dataSource: {
+          stableKey: uuid4(),
+        },
+      },
+      registry: {
+        url: new URL("/packages/npm", prismaticUrl).toString(),
+        scope: "@component-manifests",
+      },
+    };
+
+    const templateFiles = [
+      path.join("assets", "icon.png"),
+      path.join("src", "index.ts"),
+      path.join("src", "client.ts"),
+      path.join("src", "flows.ts"),
+      path.join("src", "flows.test.ts"),
+      path.join("src", "configPages.ts"),
+      path.join("src", "componentRegistry.ts"),
+      path.join(".spectral", "index.ts"),
+      path.join(".vscode", "extensions.json"),
+      path.join(".vscode", "settings.json"),
+      ".npmrc",
+      "jest.config.js",
+      "package.json",
+      "tsconfig.json",
+      "webpack.config.js",
+    ];
+
+    await Promise.all([
+      ...templateFiles.map((file) =>
+        template(
+          path.join("integration", file.endsWith("icon.png") ? file : `${file}.ejs`),
+          file,
+          context,
+        ),
+      ),
+    ]);
+
+    await updatePackageJson({
+      path: "package.json",
+      scripts: {
+        build: "webpack",
+        import: "npm run build && prism integrations:import",
+        test: "jest",
+        lint: "eslint --ext .ts .",
+      },
+      eslintConfig: {
+        root: true,
+        extends: ["@prismatic-io/eslint-config-spectral"],
+      },
+      dependencies: {
+        "@prismatic-io/spectral": "*",
+      },
+      devDependencies,
+    });
+
+    this.log(`
+"${name}" is ready for development.
+To install dependencies, run either "npm install" or "yarn install"
+To run unit tests for the integration, run "npm run test" or "yarn test"
+To build the integration, run "npm run build" or "yarn build"
+To import the integration, run "prism integrations:import"
+
+For documentation on writing code-native integrations, visit https://prismatic.io/docs/code-native-integrations/
+    `);
   }
 }
