+++ src/utils/devDependencies.ts
@@ -1,0 +1,16 @@
+/**
+ * Dev dependencies for a newly-generated code-native integration or custom connector
+ */
+
+export const devDependencies = {
+  "@prismatic-io/eslint-config-spectral": "2.1.0",
+  "@types/jest": "29.5.14",
+  "copy-webpack-plugin": "13.0.0",
+  eslint: "9.24.0",
+  jest: "29.7.0",
+  "ts-jest": "29.3.2",
+  "ts-loader": "9.5.2",
+  typescript: "5.8.3",
+  webpack: "5.99.5",
+  "webpack-cli": "6.0.1",
+};
