--- templates/connection/basic.ts.ejs
@@ -1,27 +1,0 @@
-import { connection } from "@prismatic-io/spectral";
-
-export const <%= connection.key %> = connection({
-  key: "<%= connection.key %>",
-  display: {
-    label: "<%= connection.label %>",
-    description: "<%= connection.comments %>",
-  },
-  inputs: {
-    username: {
-      label: "Username",
-      placeholder: "Username",
-      type: "string",
-      required: true,
-      comments: "Username for my Connection",
-    },
-    password: {
-      label: "Password",
-      placeholder: "Password",
-      type: "password",
-      required: true,
-      comments: "Password for my Connection",
-    },
-  },
-});
-
-export default [<%= connection.key %>];
