+++ templates/integration/src/client.ts.ejs
@@ -1,10 +1,27 @@
-import type { Connection } from "@prismatic-io/spectral";
+/**
+ * Your flows and data sources likely connect to third-party APIs.
+ * In this file, you will see an example of how to set up a reusable,
+ * authenticated HTTP client that flows and data sources can use.
+ *
+ * The createAcmeClient receives a connection that was created
+ * when a customer completed the configuration wizard. The API key
+ * and base URL from that connection are used to create the HTTP client.
+ *
+ * For more information about reusable HTTP clients, see
+ * https://prismatic.io/docs/custom-connectors/get-started/wrap-an-api/#creating-a-shared-http-client
+ */
 
-export const createClient = (connection: Connection) => {
-  // Create a client using the provided Connection for the
-  // service you're consuming from this Integration.
-  return {
-    call: async (name: unknown) =>
-      Promise.resolve(`Hello, ${name} using connection ${connection.key}`),
-  };
-};
+import { type Connection, util } from "@prismatic-io/spectral";
+import { createClient } from "@prismatic-io/spectral/dist/clients/http";
+
+export function createAcmeClient(acmeConnection: Connection) {
+  const { apiKey, baseUrl } = acmeConnection.fields;
+
+  return createClient({
+    baseUrl: util.types.toString(baseUrl),
+    headers: {
+      Accept: "application/json",
+      Authorization: `Bearer ${apiKey}`,
+    },
+  });
+}
