+++ templates/integration/src/componentRegistry.ts.ejs
@@ -1,5 +1,12 @@
+/**
+ * Your code-native integration can invoke existing connectors.
+ * This is where you declare which connectors your code-native
+ * integration uses. For more information, see
+ * https://prismatic.io/docs/integrations/code-native/existing-components/
+ */
+
 import { componentManifests } from "@prismatic-io/spectral";
 
 export const componentRegistry = componentManifests({
-  // this is where you'll register your component manifests
+  // This is where you'll register your component manifests
 });
